Added Display for stats as well as other small changes
Created AppUsageBubble(), it creates a bubble with open statistics that is displayed in rows of 2 on the default page of the app. Also added many small changes such as changes to spelling and colors in the app. Added a button that will help you add a app, to be implemented more latter.